# Speed up deserialization of large vectors by a factor of ten